+++ packages/expo-auth-session/src/__tests__/AuthRequest-test.ts
@@ -1,5 +1,5 @@
 import { AuthRequest } from '../AuthRequest';
-import { CodeChallengeMethod } from '../AuthRequest.types';
+import { CodeChallengeMethod, Prompt } from '../AuthRequest.types';
 import { buildQueryString, getQueryParams } from '../QueryParams';
 
 jest.mock('expo-random', () => ({
@@ -128,7 +128,7 @@ it(`can override the code_challenge with extraParams`, async () => {
     extraParams: { code_challenge: 'custom-value' },
   });
 
-  const url = await request.buildUrlAsync(mockDiscovery);
+  const url = await request.makeAuthUrlAsync(mockDiscovery);
   expect(getQueryParams(url).params.code_challenge).toBe('custom-value');
 });
 
@@ -140,11 +140,13 @@ it(`loads an auth request`, async () => {
     scopes: [],
     redirectUri: 'foo://bar',
     usePKCE: true,
+    prompt: Prompt.SelectAccount,
     extraParams: { code_challenge: 'custom-value' },
   });
-  await request.buildUrlAsync(mockDiscovery);
+  await request.makeAuthUrlAsync(mockDiscovery);
 
   expect(request.url).toMatch(/https:\/\/demo.io/);
+  expect(request.url).toContain('prompt=select_account');
   expect(request.codeVerifier).toBeDefined();
   expect(typeof request.state).toBe('string');
 });