+++ packages/expo-constants/src/Constants.types.ts
@@ -36,6 +36,7 @@ export interface AppManifest {
   version?: string;
   /** Published Apps Only */
   revisionId?: string;
+  releaseChannel?: string;
   orientation?: string;
   primaryColor?: string;
   icon?: string;