+++ packages/expo-splash-screen-command/bin/configureAndroidSplashScreen.d.ts
@@ -1,6 +1,0 @@
-import { ResizeMode } from './constants';
-export default function configureAndroidSplashScreen({ imagePath, resizeMode, backgroundColor, }: {
-    imagePath?: string;
-    resizeMode: ResizeMode;
-    backgroundColor: string;
-}): Promise<void>;