+++ packages/expo-updates/src/Updates.ts
@@ -13,8 +13,10 @@ import {
 
 export * from './Updates.types';
 
+export const releaseChannel: string = ExpoUpdates.releaseChannel ?? 'default';
 export const localAssets: LocalAssets = ExpoUpdates.localAssets ?? {};
 export const isEmergencyLaunch: boolean = ExpoUpdates.isEmergencyLaunch || false;
+export const isUsingEmbeddedAssets: boolean = ExpoUpdates.isUsingEmbeddedAssets || false;
 
 let _manifest = ExpoUpdates.manifest;
 if (ExpoUpdates.manifestString) {